+++ CHANGELOG.md
@@ -2,10 +2,30 @@
 All notable changes to this project will be documented in this file. 
 
 ## New Version
-### Added
-* [SER-275](https://oneacrefund.atlassian.net/browse/SER-275) As a farmer I want to be able to report seed germination issue through USSD so that it can be investigated and I can get my refund
+
 * [SER-330](https://oneacrefund.atlassian.net/browse/SER-330) Configure November Chicken confirmations - changed the chicken eligibility calculations and updated the cap
 
+
+### v1.0.9
+* [SER-323](https://oneacrefund.atlassian.net/browse/SER-323) RW - Rearrange the products in the system
+* [SER-322](https://oneacrefund.atlassian.net/browse/SER-322) RW - Add the maximum order per product in the system
+* [SER-318](https://oneacrefund.atlassian.net/browse/SER-318) Duka registration is not sending account number to elk as expected
+* [SER-277](https://oneacrefund.atlassian.net/browse/SER-277) As a OAF non client, I want to be able to access CE services through the USSD system so that I can report issues.
+* [SER-275](https://oneacrefund.atlassian.net/browse/SER-275) As a farmer I want to be able to report seed germination issue through USSD so that it can be investigated and I can get my refund
+
+## v1.0.8
+### Fixed
+* [SER-331](https://oneacrefund.atlassian.net/browse/SER-331) Duka-client sessions failing
+
+## v1.0.7
+### Fixed
+* [SER-328](https://oneacrefund.atlassian.net/browse/SER-328) USSD Failing when group size is too big
+
+## v1.0.6
+### Fixed
+* [SER-315](https://oneacrefund.atlassian.net/browse/SER-315) USSD sessions failing on API request
+
+## v1.0.5
 ### Removed
 * [SER-312](https://oneacrefund.atlassian.net/browse/SER-312) Remove all trainings except soil trainings and maize recommendation